pyproject.toml
[project]
name = "ad-campaign-analytics"
version = "0.1.0"
description = "Ad Campaign Analytics with Airflow and dbt"
requires-python = ">=3.10"
dependencies = [
    "dbt-core==1.10.11",
    "dbt-bigquery==1.10.0",
    "google-cloud-bigquery>=3.0.0",
    "requests>=2.31.0",
    "kaggle>=1.5.16",
    "pandas>=2.0.0",
    "kagglehub>=0.2.0",
    "pyyaml>=6.0.0"
]